Demonstrate that RT #121360 has been resolved.
Bug report filed by James Avera argued that "File::Spec->abs2rel($path, $base)
is supposed to allow $path and/or $base to be relative to the current working
directory, and the pod says that if either are relative, they are converted to
absolute using 'rel2abs()'."

The cases reported as failing were subsequently reported as passing.  Adding a
test file to confirm this.  TODO:  Adapt test for VMS.

For: RT #121360