Force barename base to be a directory in File::Spec::VMS:abs2rel.
The docs say that the filename portion of base is ignored, but they
don't specify what happens when base is a single component without
directory syntax, leaving ambiguous whether it's a file or a Unix-
style directory spec.  Let's default to the latter for greatest
consistency with what happens elsewhere.